rename module to avoid collision
